Revert "feat: Not Found Component (#366)"
This reverts commit cb1a62129f19efaeab645b314c834e501fd23ac6.
