+++ src/pages/booklist/containers/BooklistContainer.tsx
@@ -1,15 +1,15 @@
 import { useEffect } from 'react';
 
-import { Box } from '@chakra-ui/layout';
+import { Box, Container, Divider, Heading } from '@chakra-ui/layout';
 import { Center, Spinner, Text, VStack } from '@chakra-ui/react';
 import { useParams } from 'react-router';
 
 import { Term } from 'lib/fetchers';
+import { useDarkMode } from 'lib/hooks/useDarkMode';
 import { logEvent } from 'lib/utils/logEvent';
 import { getReadableTerm } from 'lib/utils/terms';
 
 import { Page } from 'common/layouts/Page';
-import { NotFound } from 'common/notFound/NotFound';
 
 import { BooklistHeading } from '../components/BooklistHeading';
 import { TextbookCard } from '../components/TextbookCard';
@@ -17,7 +17,10 @@ import { useTextbooks } from '../hooks/useTextbooks';
 
 export function BooklistContainer(): JSX.Element | null {
   const { term } = useParams();
+  const mode = useDarkMode();
+
   const textbooks = useTextbooks(term as Term);
+
   useEffect(() => {
     logEvent('textbooks_view', { term });
   }, [term]);
@@ -48,13 +51,18 @@ export function BooklistContainer(): JSX.Element | null {
               .map(({ sections, subject, code }) => {
                 return <TextbookCard subject={subject} code={code} sections={sections} />;
               })
-          ) : textbooks.textbookInfo.filter((textbook) => textbook && textbook.term === term).length <= 0 &&
-            textbooks.textbookInfo.length <= 0 ? (
-            <NotFound term={term} timetable>
-              Unable to find saved courses from your timetable for
-            </NotFound>
           ) : (
-            <NotFound term={term}>No textbooks found for your saved courses in </NotFound>
+            <>
+              <Divider my="4" />
+              <Container alignItems="center" maxW="container.xl">
+                <Heading size="md" color={mode('gray', 'dark.header')}>
+                  Unable to find saved courses or textbooks for{' '}
+                  <Text as="span" color={mode('black', 'white')}>
+                    {getReadableTerm(term)}
+                  </Text>
+                </Heading>
+              </Container>
+            </>
           )}
         </Box>
         {textbooks.status === 'loaded' && textbooks.textbookInfo.length > 0 && (
