+++ src/pages/registration/containers/RegistrationContainer.tsx
@@ -1,11 +1,12 @@
-import { Center, Spinner, Box } from '@chakra-ui/react';
+import { Box, Container, Divider, Heading } from '@chakra-ui/layout';
+import { Center, Spinner, Text } from '@chakra-ui/react';
 import { useParams } from 'react-router';
 
+import { useDarkMode } from 'lib/hooks/useDarkMode';
 import { useSavedCourses } from 'lib/hooks/useSavedCourses';
 import { getReadableTerm } from 'lib/utils/terms';
 
 import { Page } from 'common/layouts/Page';
-import { NotFound } from 'common/notFound/NotFound';
 
 import { RegistrationHeading } from '../components/RegistrationHeading';
 
@@ -14,6 +15,7 @@ import { CourseContainer } from './CourseContainer';
 export function RegistrationContainer(): JSX.Element | null {
   const { term } = useParams();
   const { courses } = useSavedCourses();
+  const mode = useDarkMode();
 
   // to avoid erroring out if term is not provided in URL
   // term is eventually filled in but need to avoid initial error
@@ -37,9 +39,17 @@ export function RegistrationContainer(): JSX.Element | null {
               return <CourseContainer course={course} />;
             })
         ) : (
-          <NotFound term={term} timetable>
-            Unable to find saved courses from your timetable for
-          </NotFound>
+          <>
+            <Divider my="4" />
+            <Container alignItems="center" maxW="container.xl">
+              <Heading size="md" color={mode('gray', 'dark.header')}>
+                Unable to find saved courses for{' '}
+                <Text as="span" color={mode('black', 'white')}>
+                  {getReadableTerm(term)}
+                </Text>
+              </Heading>
+            </Container>
+          </>
         )}
       </Box>
     </Page>
